requirements.txt
ipython
omegaconf
pydub
PyYAML
torch
torchaudio
pyrogram
python-dotenv
validators
openai-whisper
yt_dlp
moviepy
pysrt
num2words
transliterate
gradio_client
torch
torchvision
torchaudio